RE: rename sequence helper
